requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2==2.9.10
graphene-django==3.2.2
channels==4.2.0
